Make module compatible for import
Added two boolean options to the renyi2usm function to make plotting the renyi values optional and making a deepcopy of the cgr coordinates optional.

Moved the circular import statement to section of script executed only within the '__main__' namespace.

Suppressed all the print statements in the usm_make module.